Conflicts:
	nts.uk/web/nts.uk.pr.web/src/main/webapp/view/qmm/008/0/common/service.js
	nts.uk/web/nts.uk.pr.web/src/main/webapp/view/qmm/008/0/common/service.js.map
	nts.uk/web/nts.uk.pr.web/src/main/webapp/view/qmm/008/h/qmm008.h.service.js.map
	nts.uk/web/nts.uk.pr.web/src/main/webapp/view/qmm/008/h/qmm008.h.vm.js
	nts.uk/web/nts.uk.pr.web/src/main/webapp/view/qmm/008/h/qmm008.h.vm.js.map
	nts.uk/web/nts.uk.pr.web/src/main/webapp/view/qmm/008/i/qmm008.i.service.js.map
	nts.uk/web/nts.uk.pr.web/src/main/webapp/view/qmm/008/i/qmm008.i.vm.js
	nts.uk/web/nts.uk.pr.web/src/main/webapp/view/qmm/008/i/qmm008.i.vm.js.map

+++ nts.uk/ctx/pr/core/nts.uk.ctx.pr.core.infra/src/main/java/nts/uk/ctx/pr/core/infra/entity/rule/employment/processing/yearmonth/QpdmtPayday.java
@@ -1,6 +1,7 @@
 package nts.uk.ctx.pr.core.infra.entity.rule.employment.processing.yearmonth;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import javax.persistence.Basic;
 import javax.persistence.Column;
@@ -61,7 +62,7 @@ public class QpdmtPayday extends TableEntity implements Serializable {
 
 	@Basic(optional = false)
 	@Column(name = "NEEDED_WORK_DAY")
-	public int neededWorkDay;
+	public BigDecimal neededWorkDay;
 
 	@Basic(optional = false)
 	@Column(name = "ACCOUNTING_CLOSING")
